# tuning/internal: optimise unnecessary allocations