Fixes: use preallocated arrays where possible.

Benchmark Diff (n=64):
```
$ benchstat before.bench after.bench
goos: linux
goarch: amd64
pkg: snqk.dev/slog/meld
cpu: Intel(R) Core(TM) i7-5960X CPU @ 3.00GHz
              │ before.bench │         after.bench         │
              │    sec/op    │   sec/op     vs base        │
MeldLogger       6.709µ ± 1%   6.194µ ± 0%   -7.68% (n=64)
MeldLogger-2     3.873µ ± 0%   3.596µ ± 0%   -7.16% (n=64)
MeldLogger-4     2.167µ ± 0%   1.954µ ± 0%   -9.83% (n=64)
MeldLogger-8     2.373µ ± 1%   2.083µ ± 0%  -12.22% (n=64)
MeldLogger-16    2.998µ ± 1%   2.568µ ± 1%  -14.35% (n=64)
geomean          3.315µ        2.974µ       -10.29%

              │ before.bench │         after.bench          │
              │     B/op     │     B/op      vs base        │
MeldLogger      3.141Ki ± 0%   2.672Ki ± 0%  -14.93% (n=64)
MeldLogger-2    3.141Ki ± 0%   2.672Ki ± 0%  -14.93% (n=64)
MeldLogger-4    3.143Ki ± 0%   2.673Ki ± 0%  -14.95% (n=64)
MeldLogger-8    3.147Ki ± 0%   2.678Ki ± 0%  -14.92% (n=64)
MeldLogger-16   3.157Ki ± 0%   2.686Ki ± 0%  -14.94% (n=64)
geomean         3.146Ki        2.676Ki       -14.93%

              │ before.bench │        after.bench        │
              │  allocs/op   │ allocs/op   vs base       │
MeldLogger        73.00 ± 0%   70.00 ± 0%  -4.11% (n=64)
MeldLogger-2      73.00 ± 0%   70.00 ± 0%  -4.11% (n=64)
MeldLogger-4      73.00 ± 0%   70.00 ± 0%  -4.11% (n=64)
MeldLogger-8      73.00 ± 0%   70.00 ± 0%  -4.11% (n=64)
MeldLogger-16     73.00 ± 0%   70.00 ± 0%  -4.11% (n=64)
geomean           73.00        70.00       -4.11%
```